minor code refctor
git-svn-id: http://svn.hive.pt/hive-main/pt.hive.colony/trunk@11256 89aa9a68-8201-4d56-8cc5-2cc230aca7a0

+++ pt.hive.colony.plugins.misc/src/colony/plugins/misc/scheduler/scheduler_system.py
@@ -436,7 +436,7 @@ class Scheduler:
 
         # calculates the number of iteration to be used
         # from the sleep step
-        iterations = sleep_time / self.sleep_step
+        iterations = int(sleep_time / self.sleep_step)
 
         # calculates the extra sleep time from the sleep
         # step modulus
